Don't conditionally write ctl scripts
Doing so adds a newline before the #!/bin/bash. When running the ctl
script with bash (such as sudo su;
/var/vcap/jobs/cluster_health_logger/bin/cluster_health_logger_ctl
start), it will work fine. However, if you execute the file (for
instance, with `sudo -u root /var/vcap/jobs/cluster_health_logger/bin/cluster_health_logger_ctl
start` or with monit start cluster_health_logger), the newline before
the #! will cause problems.

We conditionally write the monit file based on the pxc_enabled variable,
so it is okay to have the ctl script always written. It just won't be
called in some circumstances. We fixed this for the
cluster-health-logger and gra-log-purger jobs.

* Also rename cluster_health_logger process to cluster-health-logger for
consistancy with other jobs

[#158443828]

Signed-off-by: Nitya Dhanushkodi <ndhanushkodi@pivotal.io>