Implement chunking for GRA log directory
This is a defensive change to prevent any memory issues if for whatever
reason the GRA log directory has an insane number of GRA logs to purge.
This commits reads the directory at a 1024 file interval. Also,
consolidated the finding and deleting of the GRA log functions into one.
Thus, we had to remove some unit tests which were redundant given our
integration test covers the cases. We also wrote a few more tests to
provide more coverage around user input and not allowing directories to
be deleted.

[#163441056]

Co-authored-by: Chris Tarazi <ctarazi@pivotal.io>
Signed-off-by: Andrew Garner <agarner@pivotal.io>