Bump pxc to 5.7.27
Updated the packaging script to not hardcode the pxc version. Instead we
glob for the source tarball and glob for the source directory after we
extract the tarball so that we can merely bump the pxc blob in the
future without also needing to bump a hardcoded version number in the
packaging script.

[#168584227]

Signed-off-by: James Palmer <jpalmer@pivotal.io>