Toggled default for relay_log_recovery
In a replication failover scenario, we need to wait for the relay logs to
be applied on the follower, BEFORE promoting it to read-writable. This
involves a restart of the node, and relay_log_recovery was set to 1,
which meant we were blowing away relay_logs, which could be potential
loss of data.

So toggling the value, so that we have the relay logs, to try to
minimize data loss when failing over.

[#164675958]

Signed-off-by: Jamie Monserrate <jmonserrate@pivotal.io>