Update counterfeiter usage in migrate-to-pxc
Elsewhere we have change the way we run "go:generate" to use the new
counterfeiter v6 utility and run counterfeiter through "go run" from the
dependency pinned in the module's go.mod.

Updated migrate-to-pxc and removed globally installing counterfeiter in
the top level scripts/test-unit as it is no longer required.