switchboard: more thorough fix to avoid race condition
The "useTLSForAgent" flag that was added can be consumed in a go routine
that we do not ensure is shutdown between tests.

Ensuring a shutdown is non-trivial without significant changes. For now
a quick fix in the test is to copy the flag before it is consumed in any
path invoked by a go routine.