galera-healthcheck: allow config tests to run concurrently
Previously hardcoded port 8080 in tests and asserted that the
constructed "NetworkListener" was configured with or without TLS per the
configuration file by testing a TLS/plaintext connection to the
listener. This could lead to two listeners attempting to listen on the
same port when running tests concurrently (e.g. ginkgo -p or similar)

Now a random port is set in the config so that these tests can run
concurrently.