specs: stop hardcoding /usr/local/bin/bosh
This is only used in a single test, bosh_wait_scripts, and assume the
bosh cli has been installed in /usr/local/bin/bosh.  This is not
guaranteed to be the location of the BOSH cli and we should be free to
use other CI image that may install bosh elsewhere.

Now the `ExecuteBosh` helper assume "bosh" is in PATH and does not
require a hardcoded absolute path to the bosh cli.