Support pxc-8.0 blobs on jammy
Percona compiles their binary distribution against libldap-2.4 but fails
to vendor this library. When attempting to run these binaries on Jammy
(or any other system without libldap_r-2.4.so) the mysqld will fail with
an error like so:

> ./bin/mysqld: error while loading shared libraries:
libldap_r-2.4.so.2: cannot open shared object file: No such file or directory

This change applies a minimal hack to ensure libldap is resolvable so
mysqld should run. The underlying ldap functionality may be indeed be
broken but mysqld seems to run without issue.  This is a short-term fix.
Long-term the packaging for PXC 8.0 will be updated to avoid these kinds
of issues.

[#182536919](https://www.pivotaltracker.com/story/show/182536919)