cleanup unused blobs and packages
Remove all the PXC 5.7 bits. For upgrades and clustering the needed
components are provided by the PXC 8.0 binary distribution (e.g.
xtrabackup 2.4 + 8.0 as well as libgalera)

This avoids wasting time and resources in CI compiling bits we are not
using.  In the future we may bring some of these source bits back in a
different form if we move away from the Percona binary distribution.

[#182537006](https://www.pivotaltracker.com/story/show/182537006)

Authored-by: Andrew Garner <garnera@vmware.com>