switchboard: allow binding to localhost
Introduce a `BindAddress` property to the proxy configuration.  This
defaults to "empty" / binding to all interfaces and replaces the
previously hardcoded "0.0.0.0:$port" or ":$port" syntax we previously
used.

The main driver for this change is avoiding a slew of popups on MacOS
when an unauthorized service attempts to accept incoming network
connections outside of loopback.  This makes it easier to run the
switchboard tests locally.

[#178563165](https://www.pivotaltracker.com/story/show/178563165)

Authored-by: Andrew Garner <garnera@vmware.com>