get-sequence-number: improve script
Backfilled integration test for get-sequence-number

Rewrote the get-sequence-number script to fix two problems at once:

- Script will run with galera-agent tls enabled because it no longer
  interacts with galera-agent

- Script will run when galera-agent is offline because it no longer
  interacts with galera-agent

This script is useful to determine the most up-to-date node which is
most often done for manual bootstrapping when galera-agent is offline.

Decoupling the behavior of this script from the requirement of
galera-agent behing online make this more reliable for human
consumption.

[#182537255](https://www.pivotaltracker.com/story/show/182537255)

Co-authored-by: Andrew Garner <garnera@vmware.com>
Co-authored-by: Kyle Ong <kyleo@vmware.com>