galera-init: remove seeded_databases / seeded_users and poststartSQL features
Seeded users is implemented as part of the pxc-mysql job, but no longer
the responsibility of galera-init.

The functionality of seeded_databases can now be specified as
seeded_users.  This is currently a breaking change, but with a
functional path forward for a user.

PostStartSQL was never supported by the pxc-mysql job, but there was
coding supporting the feature lingering in galera-init.  This has been
removed as dead code.

[#182920141](https://www.pivotaltracker.com/story/show/182920141)

Authored-by: Andrew Garner <garnera@vmware.com>