smoke-tests: handle slow cluster writes
The new smoke-tests implementation was pulld in from the pxc-8.0 branch
which was relying on a PXC 8.0 feature to set wsrep_sync_wait for only
the specific read query being issued.  That feature is not supported and
ignored on PXC 5.7.

This change set the wsrep_sync_wait=1 session variable for the
smoke-tests connection so all reads will be synchronous and avoid flake
due to slow writes.