pxc-mysql: automatically generate seeded_users for system components
For galera-agent and cluster-health-logger the pxc-mysql job will now
consume a link and generate an appropriate seeded_users entry for those
users so that a consumer of pxc-release is not required to configure a
specific user for galera-agent or cluster-health-logger.

smoke-tests is not automated in this way.  Since that errand is largely
used for testing, the burden on the consumer to configure a specific
smoke-tests username/password is not considered unreasonable and avoids
yet another user to track.

galera-agent and cluster-health-logger are maintained largely for
historical reasons and ease adoption of pxc/1.0

[#183065535](https://www.pivotaltracker.com/story/show/183065535)

Co-authored-by: Andrew Garner <garnera@vmware.com>
Co-authored-by: Colin Shield <cshield@vmware.com>
Co-authored-by: Kim Bassett <kbassett@vmware.com>