pxc-mysql: set pxc-maint-transition-period to zero seconds
During a rolling upgrade from PXC 5.7 to PXC 8.0 the 8.0 instances
remain in a read-only state until the last PXC 5.7 instance has exited
the cluster. The proxy job will not route traffic to PXC 8.0 instance
until the PXC 5.7 has shutdown, but with the default
pxc-maint-transition-period of 10 seconds PXC 5.7 takes extra time to
actually shutdown.

This commit sets this property to zero so a PXC 5.7 will shutdown more
quickly.

[#182920179](https://www.pivotaltracker.com/story/show/182920179)

Co-authored-by: Andrew Garner <garnera@vmware.com>
Co-authored-by: Kyle Ong <kyleo@vmware.com>