pxc-mysql: conditionally perform pxc-5.7 crash recovery in pre-start
In order to avoid a failed upgrade to pxc-8.0, pre-start checks if the
data directory is still a pxc-5.7 formatted directory and will attempt
to run a crash recovery command using the pxc-5.7 distribution.

A pxc-5.7 data directory is detected by looking for the .frm file for
the mysql.user table.  This file format was removed in PXC 8.0 so if we
detect an .frm file we know the data directory has not been upgraded
yet.

[#183085049](https://www.pivotaltracker.com/story/show/183085049)

Co-authored-by: Andrew Garner <garnera@vmware.com>
Co-authored-by: Kyle Ong <kyleo@vmware.com>