galera-agent: re-add the drain script
We found during a "delete-deployment" operation, the drain script can
sometimes lead to a flake as one node may interpret another node as
being unhealthy.  The drain script does not know whether the deployment
is being deleted so cannot do anything more interesting at the moment.

This behavior seems to be cause by the galera-agent staying online
longer than the previous behavior in v0.  This means the agent will
report a MySQL instance as unhealthy and cause the drain script to stop.
If galera-agent stops _before_ mysql, the drain script will ignore the
instance.

Long-term we should convert the drain script to a pre-stop script and
avoid this unnecessary logic during a delete-deployment.  For now we are
returning the pxc v0 behavior.

This reverts commit a86be60857e4968cecaaeea582e83f9c6a6ff7cd.

[#183819906](https://www.pivotaltracker.com/story/show/183819906)