galera-init: make integration tests more portable
When building the galera-init binary on the host system, the resulting
binary may link to a libc that is incompatible with the underlying
container.  This is known to happen when building on jammy and running
the Percona images (based on a RHEL 8 image).

To avoid this, build with CGO_ENABLED=0. The galera-init binary should
not rely on any specific cgo behaviors anyway.