packages: allow parallel compilation of server and xtrabackup packages
Previously percona-xtradb-cluster-8.0 depended on the percona-xtrabackup-*
packages explicitly forcing serial compilation of xtrabackup and
percona-xtradb-cluster.

Now, percona-xtradb-cluster-8.0 relies on a symlink to
percona-xtrabackup-{2.4,8.0} and can compile in parallel to the
xtrabackup packages.

The "pxc-mysql" job already depended directly on percona-xtrabackup-8.0
and 2.4 so this is only a small optimization to speed up compilation and
reduce the compiled bosh release by ~45MiB.  In practice we expect this
to halve compilation time when > 3 bosh compilation workers are
available.