pxc-mysql: upgrade 5.7 instances to 8.0 prior to normal startup
This is a simple change that upgrades mysql with the 8.0 database server
binary when the previous data directory was detected as mysql v5.7.

This changes mitigates https://bugs.mysql.com/110702 by forcing an extra
start / stop cycle in the database before any additional schema changes
are applied which avoids this data corruption bug.

An e2e-test was added to better drive out this defect in a clustered
environment. This provides a test to verify that nodes are upgraded
correctly both locally and via Galera state transfer.  Previously, the
test that was modified was performing an upgrade test from a single node
cluster to a three node cluster which avoided some of the more hairy
upgrade cases.

[#185222532](https://www.pivotaltracker.com/story/show/185222532)

Co-authored-by: Andrew Garner <garnera@vmware.com>
Co-authored-by: Kim Bassett <kbassett@vmware.com>