Removes replace directive in switchboard
Previously this was referencing the galera-healthcheck folder, for the
purposes of importing one API response model

This caused problems in 'bump-go-dependencies' job in CI that we missed.
That job would iterate through all the projects, running `go mod` and
`go tidy`. But since the switchboard dependencies were bumped BEFORE
healthcheck, that meant the switchboard dependencies could become
out-of-date AFTER the healthcheck dependencies are bumped

This led to more downstream problems because our OSSPI jobs run
`/go/bin/go list -mod=readonly -m -json all` to determine what packages
to upload to OSM. BUT that `go list` command fails in the switchboard
because it wants to first run `go mod tidy`

This fix removes the dependency on the healthcheck project, meaning that
the order of bumping go dependencies won't matter. Switchboard will
always have the most up-to-date list of dependencies, and the OSSPI jobs
won't have any problems

[#186026895](https://www.pivotaltracker.com/story/show/186026895)

Co-authored-by: Ryan Wittrup <rwittrup@vmware.com>
Co-authored-by: Andrew Garner <garnera@vmware.com>