pxc-mysql: run mysql_upgrade workflow for mysql-5.7 instances
Conditionally run mysql_upgrade when the contents of the
mysql_upgrade_info file in the datadir does not match the current
version reported by the mysqld 5.7 binary.

This addresses an upgrade issue in Percona XtraDB Cluster 5.7.43 where a
catalog change was made to increase the column size of various tables,
most notably performance_schema.global_variables.VARIABLE_VALUE which
requires an upgrade.

Previously, the mysql_upgrade workflow had been removed as it was broken
under MySQL 8.0 and added significant complexity to the galera-init
component. Percona XtraDB Cluster 8.0 handles upgrades automatically.
Percona XtraDB Cluster 5.7 generally did not require this upgrade
workflow for patch releases until
5.7.43.

Reintroduced the upgrade workflow as a simple bash script to avoid
adding complexity to the already complex galera-init component and to
make it easier to delete MySQL 5.7 support in the future as this MySQL
version reaches its end-of-life.

[#185952849](https://www.pivotaltracker.com/story/show/185952849)