galera-init: implement wsrep-recovery when starting mysqld nodes
This addresses an old defect in pxc-release where a mysql node that
previously crashed or ungracefully terminated would restart and perform
a full state transfer from the cluster rather than incremental recovery.

Now before joining the local node to an existing cluster attempt to run
"mysqld ... --wsrep-recover" and determine the local nodes actual state.
If found, mysqld is run with the "--wsrep-start-position" option to
improve the chances of a faster incremental cluster join operation.

If no position can be found or if the logs report the zero gtid
("00000000-0000-0000-0000-000000000000:-1"), then the old behavior is
used and mysqld is started normally.  In this case the node will likely
attempt a full state transfer from the cluster.

[#185952849](https://www.pivotaltracker.com/story/show/185952849)