switchboard: log health-check failures better
Previously the raw Go stdlib http.Response value was logged - typically
a bunch of useless pointer values with the raw http status mixed in
somewhere.

Now the http status is consistently logged on failure along with the
body of the response from the status endpoint.

Errors communicating with the status endpoint are clearly recorded in
the "error" field and the previous duplicated "err" field is omitted.

Some cleanup to the implementation was done here as well:

- Ensure the http body is closed and fully read / discarded even in case
  of error. This avoids a potential resource leak in the proxy when the
  cluster is in an unhealthy state.

- Flatten some of the ugly nested logic by checking exception conditions
  up front

[#184993377](https://www.pivotaltracker.com/story/show/184993377)

Co-authored-by: Andrew Garner <garnera@vmware.com>
Co-authored-by: Kyle Ong <kyleo@vmware.com>