tests: fix jemalloc profiling test flake
Previously ran flush against the active writer, but checked for a file
on mysql/0. There was only a one-in-three chance mysql/0 was the right
node and if not, the test would fail.

Now ensure the node the flush query is run on matches the node we are
checking a file on.