Backport fix from PS-9369: Mitigates "Audit plugin leaks memory"
When a workload with long-lived connections that repeated execute
prepared statements the Percona 8.0 audit log plugin can lose track of
statements and bloat the size of an internal stack data structure. Over
time this stack doubles in size leading to excessive memory utilization.

The patch for PS-9369 mitigates this issue by tracking queries by
content rather than by pointer values avoiding "leaking" data onto a
stack that never gets cleaned up.

This is a backport of a fix expected in a future version of Percona
XtraDB Cluster 8.0. Once a version of Percona XtraDB Cluster that
includes the fix for PS-9369 is released this commit may be reverted.

[TPCF-27370](https://vmw-jira.broadcom.net/browse/TPCF-27370)

Reviewed-by: Kim Bassett <kim.bassett@broadcom.com>