e2e-tests: relax test assertion for cluster-health-logger
cluster-health-logger logs the PXC wsrep status counters and reports
whether binary logging is enabled as a delimiter separated value file.

The e2e-test for this component was looking for a header and hardcoded a
sequence of variables.  In PXC v8.0.41 new status counter were
introduced that broke the ordering assumption and caused this test to
fail.

This commit updates the test assertion to more loosely look for this
header and look for roughly valid output without hardcoding specific
status counters.