Add pxc-mysql job property to disable check for persistent storage
We ran into a problem where this check was blocking deployment when trying to
deploy cf-deployment in a bosh-lite situation with the docker-cpi running inside
of another docker container.

The device id seems to be the same in that situation even though the docker cpi
is creating separate volume mounts into the container.