spike: patch binlog_utils_udf to compare normalized paths
When Galera provisions a node via SST, some process copies the last
binary log from the donor node to preserve GTID history. With default
MySQL configuration this log name is added to the log-bin-index file as
"mysql-bin.00000N" rather than "./mysql-bin.00000N" as MySQL would
normally do this.

In practice, this doesn't cause any problems - both paths are
equivalent.  MySQL can manage, read and process this slightly oddly
named binlog entry the same as any other binlog.

However, the Percona extension binlog_utils_udf.get_gtid_set_by_binlog,
if installed, fails to read this binary log.  The reason is purely due
to bad string comparisons.   The function normalizes its input, i.e.:

> SELECT get_gtid_set_by_binlog('mysql-bin.000001')

based on the log-bin-index.  So this becomes "./mysql-bin.000001"

Later it is compared to the string "mysql-bin.000001" which is not equal
- the entry is missing a leading "./" even though the paths are
  logically equivalent.

This can happen in a few cases.  Initially we observed this behavior
after a Galera SST, but this could also happen if the log-bin-index
option is changed in the future.  In either case the directory prefix
may differ across binary logs even if binary logs are logically in the
same path on the storage device.  I.e. all these paths are equivalent
but Percona's get_gtid_set_by_binlog function will fail on some of them:

- mysql-bin.000001
- ./mysql-bin.000001
- /var/vcap/store/pxc-mysql/mysql-bin.000001

This spike patches PXC to compute the basename of both the input and
each entry in the log-bin-index prior to comparison.  This adds a little
more overhead relative to the previous implementation as a substring is
extracted from the basepath and possible some additional memory
management done.  Some microbenchmarking did not indicate any
significant runtime different on a system with 2000 binary logs reading
the first binary log in the list.