e2e-tests: test upgrades from latest EOL 5.7 to 8.0
Previously testing 0.44.0, but now testing against the last patch
release of the pxc v0 series that predated the jump to MySQL v8.0.

Unpins bpm to use latest provided by the bosh environment. bpm _must_ be
an available release in the bosh director.

Stops uploading a pxc v0 _source_ release and assumes this has been
preconfigured in the environment. This avoids recompilation during a
test run if a precompiled test asset was uploaded with an older
stemcell.