e2e-tests: record output from RemoteCommand better
Troubleshooting why a bosh ssh command is flaking sometimes in an
internal environment.  When _only_ capturing the "Stdout" column, other
troubleshooting information is lost.

Now the RemoteCommand implementation requests json output from the bosh
cli and extract the "stdout" value.  On error the entire bosh response
will be surfaced to aid in troubleshooting.