Atomically write state file using google/renameio
Discovered when running e2e-tests in a vsphere environment. The state
file ended up in an "empty" state after a vm was deleted.  It seems
likely this was caused by failing to fsync the underlying file
correctly.

Using google/renameio which handles a wide variety of corner cases for
correctly syncing to disk while providing the same WriteFile interface
as the Go stdlib.