e2e-tests: handle reconnecting to available node on scale-down
This test previously assumed that proxy/0 would be retained after a
scale down operation.

In a vsphere environment, that is not necessarily the case and bosh
sometimes retains the second proxy vm instead.

This test was connecting to a precached proxy IP, and if this assumption
does not hold then database connections fail after scale-down and the
test fails.

Now, force a reconnect to the available proxy after a scale-down
operation in order to discover the available proxy VM.