e2e-tests: remove assertion on "jeprof" utility
jeprof is a utility shipped with the jemalloc memory allocator library
to aid in analysis of memory profiles during manual debugging
operations.

This utility relies on the "objdump" binary which is a dev utility
installed on the stemcell.  In production deployments, the bosh director
is configured with "remove_dev_tools: true" and objdump is NOT available
on the VMs.

When running e2e-tests in these environments, jeprof is not available
and the test will fail.

Remove the jeprof test entirely - it is not part of the core pxc-release
functionality and ancillary behavior provided by the jemalloc runtime
dependency.