tweak innodb buffer pool autotune test
Ensure expectation is adjusted for the block alignment done by the
InnoDB storage engine.

As of MySQL v5.7 the innodb buffer pool size is aligned to the
innodb_buffer_pool_chunk_size x innodb_buffer_pool_instances.

The actual defaults for these change a bit over MySQL versions, but the
core computation should be reliable on MySQL v5.7 and later.

The innodb_buffer_pool_size_percent value used by the test when
deploying a pxc deployment is now configurable.  The default is "14" to
match the old test, but can be overriden by setting the
INNODB_BUFFER_POOL_SIZE_PERCENT environment variable. This makes it
easier to flush out incorrect test (or implementation) logic that relies
on magic values