pxc-mysql: scale wsrep_applier_threads when mysql_version=8.4
Since the option is set in a version specific section of the my.cnf, the
implementation needed to be updated to set it in a [mysqld-8.4] section.

This can be simplified in the future when mysql v5.7 support is dropped
and the legacy "wsrep-slave-threads" option no longer needs to be
supported.

[TNZ-42275](https://vmw-jira.broadcom.net/browse/TNZ-42275)

Authored-by: Andrew Garner <andrew.garner@broadcom.com>