pxc-mysql: ensure latest mysql cli is installed in /usr/local/bin/mysql
Previously the /usr/local/bin/mysql program was only updated if it did
not exist (i.e. maybe on a stemcell update) but it was NOT updated if it
already existed but was at an older version.

Now prestart always relinks /usr/local/bin/mysql to point to the latest
version. This avoids using a mysql cli from an older major.minor when
upgrading.

# Conflicts:
#	jobs/pxc-mysql/templates/pre-start.sh.erb