Allows users to be created with defined auth plugins
Adds new job spec property for setting auth plugin to support
transitioning from mysql_native_password to caching_sha2_password

The property defaults to mysql_native_password to fully support existing
consumers, but it allows manually set the plugin when adding a seeded
user / seeded database. This will allow consumers to define when they
want to make the migration

MySQL team can open a PR to TPCF to set the properties for all the users
we create for the platform, and this lets us opt out individual users if
we find an application doesn't support the new plugin yet

For consumers that do not define an explicit plugin for a given user,
this currently defaults to sticking with mysql_native_password to allow
for backwards compatibility - maybe we flip this in the next PXC
release?

Verification tests:
For the 'external' type users, who are not limited to acces via
localhost, test via properly configured MySQL connector that does not
allow native password usage

For the 'internal' type users, still test via SSH onto box to query
locally.