Uses docker cli for galera-init integration tests
Removes need for the larger package that directly wraps a docker client
to slim down our dependencies.

This does mean that running these tests requires users to have a
docker-compatible CLI configured on their PATH such as docker or podman

This also fixes a possible compatibility bug in how the galera-init
binary is built. We previously hardcoded to build for an amd64
architecture, which can lead to hard-to-decipher errors if the tests are
run on new Apple silicon

Co-authored-by: Ryan Wittrup <ryan.wittrup@broadcom.com>
Co-authored-by: Andrew Garner <andrew.garner@broadcom.com>