[PoC] implement automated cluster recovery
Add feature flag "automate_cluster_recovery" that causes pxc-mysql to
skip failures in post-start.  This allows the entire cluster to roll so
that a separate job run _after_ the mysql instances can repair the
cluster.

This relies on a simple post-start script that runs the bootstrap
errand.  This does not heal a partially offline cluster, but will be a
no-op for a healthy cluster and restore a fully crashed cluster.

This is just a Proof-of-concept (PoC) that could be improved in a few
ways:

- Instead of having a separate instance group, maybe move recovery to
  the "last" instance of the mysql instance group

  This may be tricky as it depends precisely on what order bosh deploys
  VMs.  Ideally we want to run a script on mysql/2 (i.e. the last node
  of a 3 node cluster), but bosh may sometimes run mysql/0, mysql/2,
  mysql/1 out of order (or maybe an order on some other criteria).
  Needs clarification

- Instead of blindly run the bootstrap errand do more intelligent repair
  of the cluster.

  If completely offline, bootstrap if all nodes are part of the same
  primary component and start (bootstrap) the most up-to-date instance
  first.

  If only partially offline, rejoin nodes to the cluster if they are
  part of the same primary component (can we even tell this if a cluster
  had been bootstrapped?)