proxy: add pingdb subcommand to proxy
pingdb is expected to be colocated with the proxy and connects over
localhost to the local proxy instance on the proxy port and establishes
a MySQL connection without credentials.

If a MySQL server response is detected, the proxy is routing connections
to a valid backend and pingdb succeeds.  Otherwise pingdb fails on any
other error.

Notes:
- Move the switchboard proxy main entrypoint to ./cmd/proxy/
- Adds new ./cmd/pingdb to aid in DNS healthchecks for the proxy
- Adds new internal/testing/docker package to assist in integration
  tests for the new pingdb command

[TNZ-66259](https://vmw-jira.broadcom.net/browse/TNZ-66259)

Authored-by: Kim Bassett <kim.bassett@broadcom.com>