Use env config parsing library
This simplifies some of the otherwise squirrely logic of manually
parsing environment variables with a relatively lightweight dependency.

This makes parsing / validation a little simpler and much easier to
extend or use richer data types in the future.

This commit changes the "TIMEOUT" input to be a Go duration rather than
a bare integer to allow for easier / more fine-grained configuration.