e2e-tests: Add test for bootstrapping after node recreate
Adds a more complex bootstrap scenario by forcing a cluster to lose
quorum and subsequently triggering a node recreate.  This emulates real
production scenarios where a cluster fails and an unresponsive node is
recreated.

The expectation is that after the work in the current story, "bootstrap"
can be trivially run to restore a working cluster.

As of this commit, this is a failing test because mysql fails in
pre-start and the bosh instance is left in a state that bootstrap cannot
trivially manage the mysql instances.

[TNZ-67462](https://vmw-jira.broadcom.net/browse/TNZ-67462)